-Wconversion issues
git-svn-id: https://svn.r-project.org/R/trunk@59198 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/par.c
@@ -269,7 +269,7 @@ static void Specify(const char *what, SEXP value, pGEDevDesc dd, SEXP call)
         x = asReal(value);
         posRealCheck(x, what);
         R_DEV__(cex) = 1.0; /* ! (highlevel par, i.e.  Specify2(), set x ! */
-        R_DEV__(cexbase) = (float)x;
+        R_DEV__(cexbase) = x;
     }
 
     else if (streql(what, "fg"))
@@ -413,12 +413,12 @@ static void Specify(const char *what, SEXP value, pGEDevDesc dd, SEXP call)
         R_DEV_2(layout) = FALSE;
         if (nrow > 2 || ncol > 2)
         {
-            R_DEV_2(cexbase) = (float)0.66;
+            R_DEV_2(cexbase) = 0.66;
             R_DEV_2(mex) = 1.0;
         }
         else if (nrow == 2 && ncol == 2)
         {
-            R_DEV_2(cexbase) = (float)0.83;
+            R_DEV_2(cexbase) = 0.83;
             R_DEV_2(mex) = 1.0;
         }
         else
@@ -446,12 +446,12 @@ static void Specify(const char *what, SEXP value, pGEDevDesc dd, SEXP call)
         R_DEV_2(layout) = FALSE;
         if (nrow > 2 || ncol > 2)
         {
-            R_DEV_2(cexbase) = (float)0.66;
+            R_DEV_2(cexbase) = 0.66;
             R_DEV_2(mex) = 1.0;
         }
         else if (nrow == 2 && ncol == 2)
         {
-            R_DEV_2(cexbase) = (float)0.83;
+            R_DEV_2(cexbase) = 0.83;
             R_DEV_2(mex) = 1.0;
         }
         else
@@ -1391,12 +1391,12 @@ SEXP attribute_hidden do_layout(SEXP call, SEXP op, SEXP args, SEXP env)
 
     if (nrow > 2 || ncol > 2)
     {
-        gpptr(dd)->cexbase = dpptr(dd)->cexbase = (float)0.66;
+        gpptr(dd)->cexbase = dpptr(dd)->cexbase = 0.66;
         gpptr(dd)->mex = dpptr(dd)->mex = 1.0;
     }
     else if (nrow == 2 && ncol == 2)
     {
-        gpptr(dd)->cexbase = dpptr(dd)->cexbase = (float)0.83;
+        gpptr(dd)->cexbase = dpptr(dd)->cexbase = 0.83;
         gpptr(dd)->mex = dpptr(dd)->mex = 1.0;
     }
     else
